refactor(WebAPI): Replace ICommandExecutor with IReaparrDbContext in ReaparrDbContextManager
